requirements.txt
Flask==0.10.1
Flask-Classy==0.6.10
Flask-SQLAlchemy==2.0
Flask-Script==2.0.5
Flask-Testing==0.4.2
Jinja2==2.7.3
Mako==1.0.0
MarkupSafe==0.23
SQLAlchemy==0.9.7
Werkzeug==0.9.6
alembic==0.6.7
blinker==1.3
factory-boy==2.4.1
itsdangerous==0.24
nose==1.3.4
psycopg2==2.5.4
raven==5.0.0
requests==2.4.3
uWSGI==2.0.7
wsgiref==0.1.2
yummly==0.4.0
